Added Data Models in ApplicationDbContext.cs
made GeneralCurrencyId as nullable in Customer.cs
Added models for WorkOrder , WorkOrderLabor , WorkOrderPartNumber , WorkOrderStage , WorkOrderStatus , WorkOrderType
Added a new method in CustomerRepository to get the Customer specific information.
Added a new method in EmployeeRepository to get the Employee Specific information.
Added a method for Implementattion in ICustomerRepository.
Added a method for Implementation in IEmployee.
Imported the service in app.module.ts.
Added TypeScripts models for the same Database models.

+++ QuickApp.Pro/ViewModels/CustomerViewModel.cs
@@ -18,7 +18,7 @@ namespace QuickApp.Pro.ViewModels
     public class CustomerViewModel : PasBase
     {
         public long CustomerId { get; set; }
-       
+
         public string Name { get; set; }
 
 
@@ -36,7 +36,7 @@ namespace QuickApp.Pro.ViewModels
 
         public long? CustomerClassificationId { get; set; }
 
-       // public int? IntegrationPortalId { get; set; }
+        // public int? IntegrationPortalId { get; set; }
         public string Discount { get; set; }
         public string EDIDescription { get; set; }
 
@@ -65,7 +65,7 @@ namespace QuickApp.Pro.ViewModels
 
         public string CustomerURL { get; set; }
 
- 
+
         public long? Addressid { get; set; }
 
         public bool IsCustomerAlsoVendor { get; set; }
@@ -102,7 +102,7 @@ namespace QuickApp.Pro.ViewModels
         public bool? EDI { get; set; }
 
 
-       // public string EDIDescription { get; set; }
+        // public string EDIDescription { get; set; }
 
         public bool? AllowProformaBilling { get; set; }
 
@@ -118,7 +118,7 @@ namespace QuickApp.Pro.ViewModels
 
         public string PrimarySalesPersonId { get; set; }
 
-       public string SecondarySalesPersonName { get; set; }
+        public string SecondarySalesPersonName { get; set; }
 
         public string SecondarySalesPersonId { get; set; }
 
@@ -143,7 +143,7 @@ namespace QuickApp.Pro.ViewModels
 
         public bool? ExportLicenseRequired { get; set; }
 
-       // public string DiscontValue { get; set; }
+        // public string DiscontValue { get; set; }
 
         public long? DiscountId { get; set; }
 
@@ -160,7 +160,7 @@ namespace QuickApp.Pro.ViewModels
 
         public Int32? MarkUpPercentageId { get; set; }
 
-        public Int64? GeneralCurrencyId { get; set; }
+        public Int32? GeneralCurrencyId { get; set; }
         public bool? IsAddressForBillingAndShipping { get; set; }
         public virtual MasterCompany MasterCompany { get; set; }
